Debug serve_on_available_port (must do check)
Previously, the serve_on_available_port decorator assumed that if you
tried to run either front- or back-end on a port that was already in
use, the attempt would throw an error. The function would then catch the
error, and try a different port.

While this worked for the front-end, as http.server does throw an error
if it finds the port in use, it did not work with Flask, which just
prints a message that the port is busy, and then gracefully stops trying
to run the app, without throwing any errors.

So now we explicitly use socket to check whether the port is busy.