Add logging to run.py script
Logging should capture process ids for parent and process, and log all
output to stdout, as well as recording runtime.
I found the tricky part was setting things up so that the stdout from
both processes was collected and saved, while also still running it to
the terminal while in progress.