Add dynamically generated RESTful API resources
Given a defined database model, we now have a function which makes it
trivial to create an object resource and a list resource which allow the
user to get, update, delete, list, and add new records -- ie. the full
set of CRUD methods.
And we have Flask-Praetorian authentication, which allows any and only
authenticated users to get, list and add records, and only admins to
update and delete records.