Avoid Netlify setting up Python env
If Netlify sees the pyproject.toml file in the root, it panics; because
it thinks it needs to install the repo as a Python package, but it
doesn't know what to do with pyproject.toml files and so gives up with
the excuse that it only handles legacy setup.py setups.

So we hide the pyproject.toml, setup.cfg and requirements.txt files in
the py_setup folder, and it seems to work.