Add CRUD functionality to frontend
On login, the user is now given a dropdown menu allowing him to select
one of the available resources. When selected, a GET request is made for
the resource list, generating a table with edit and delete buttons for
each row. And, so long as the resource is not 'users' (which can only be
created directly by signup), there is also an add button. All
authenticated users should be able to get resource lists and add new
resources; only admins should be able to update and delete.