Add freeze of pip packages
The list is generated by creating a new conda env with the
./create_env.sh script at repo root, then moving into 'py_setup' and
doing 'pip install -e .', which installs the local
package editably, and its requirements; and then doing
'pip freeze > dependencies.freeze'--but removing the repo installation.
In particular, note that the frozen list of dependencies does not
include the other requirements for local development (contained in
requirements.txt), but only the package dependencies (ie. it contains
details of specific versions of the packages listed -- explicitly and
implicitly--in setup.cfg).