Add git API resource
The main motivation for this is to make it possible to remotely trigger
the API server to pull the latest changes from Github: this is done with
a POST request, since that is the only option for Github webhooks. But
the other methods also do things, which might or might not be useful.